Compile UpdateOptions_ServerCertificateContextProvided test on desktop only
The UpdateOptions_ServerCertificateContextProvided_DoesNotDisposeCallerContext test exercises SslStreamCertificateContext.Create with intermediate certificates, which pulls in a generic Dictionary instantiation whose IL body is stripped from the Apple-mobile CoreCLR ReadyToRun composite image, crashing System.Net.Security.Unit.Tests at startup before test discovery.

Move the single failing test into a new SslAuthenticationOptionsTests.Desktop.cs partial class file and register it via an MSBuild Compile condition limited to desktop platforms (windows/unix/osx). The rest of the test suite keeps running on mobile.

Fixes #131922
